script/1080/pic1080/requirements.txt
pyautogui==0.9.53
opencv-python==4.5.5.64
numpy==1.21.2
psutil==5.8.0
keyboard==0.13.5
pystray==0.17.0
Pillow==8.4.0
